test(api): finished user tests and linted

+++ app/tasks.py
@@ -8,25 +8,29 @@ from flask import render_template, url_for
 from bs4 import BeautifulSoup
 import re
 
-
-from app import create_app, db, s3, bucket
+import app
+from app import create_app, db, s3, bucket, CustomLogger
 from app.export import get_zip
 from app.email import send_email
 from app.main.ebooks import epubTitle, epubConverted
 from app.main.pdf import importPDF
 from app.models import Task, Article, Highlight, Tag, User, Event
 
+logger = CustomLogger('Tasks')
 
-app = create_app()
-app.app_context().push()
+if not os.environ.get('testing'):
+    app = create_app()
+    app.app_context().push()
 
-def delete_user(u):
+def delete_user(id):
+    u = User.query.filter_by(id=id).first()
     highlights=u.highlights.all()
     topics = u.topics.all()
     articles = u.articles.all()
     tags = u.tags.all()
     senders = u.approved_senders.all()
     comms = u.comms
+    
     for h in highlights:
         db.session.execute(f'DELETE from highlights_topics where highlight_id={h.id}')
         db.session.execute(f'DELETE from tags_highlights where highlight_id={h.id}')
@@ -43,8 +47,9 @@ def delete_user(u):
         db.session.delete(a)
     for s in senders:
         db.session.delete(s)
-    db.session.delete(comms)
-    #db.session.delete(u)
+    if comms:
+        db.session.delete(comms)
+    
     u.email=None
     u.goog_id=None
     u.firstname = None
@@ -59,12 +64,12 @@ def _set_task_progress(progress):
     job = get_current_job()
   
     if job:
-        print(f'Job: {job}')
+        logger.info(f'Job: {job}')
         job.meta['progress'] = progress
         job.save_meta()
-        print(f'task id: {job.get_id()}')
+        logger.info(f'task id: {job.get_id()}')
         task = Task.query.get(job.get_id())
-        print(f'Task: {task}')
+        logger.info(f'Task: {task}')
         task.user.add_notification('task_progress',
                                    {'task_id': job.get_id(),
                                      'progress': progress })
@@ -76,9 +81,9 @@ def _set_task_progress(progress):
 def account_export(uid, ext, delete=False):
     try:
         _set_task_progress(0)
-        user = User.query.get(uid)
+        user = User.query.filter_by(id=uid).first()
         if ext != 'none':
-            print('exporting')
+            logger.info('exporting')
             if os.environ.get('DEV'):
                 az_path_base = f'staging/{user.id}/exports/'
             else:
@@ -91,42 +96,20 @@ def account_export(uid, ext, delete=False):
                 Filename=filename,
                 Key=az_path
                 )
-            location = s3.get_bucket_location(Bucket=bucket)['LocationConstraint']
-            #url = "https://s3-%s.amazonaws.com/%s/%s" % (location, bucket, az_path)
             url = s3.generate_presigned_url('get_object', Params = {'Bucket': bucket, 'Key': az_path}, ExpiresIn = 604800)
 
             delete_date = (datetime.today()+ timedelta(days=7)).strftime("%B %d, %Y")
-            print(f'sending email - [Lurnby] Your exported data for user: {user.id}')
+            logger.info(f'sending email - [Lurnby] Your exported data for user: {user.id}')
             send_email('[Lurnby] Your exported data',
                     sender=app.config['ADMINS'][0], recipients=[user.email],
                     text_body=render_template('email/export_highlights.txt', url=url, user=user, delete_date=delete_date),
                     html_body=render_template('email/export_highlights.html', url=url, user=user, delete_date=delete_date),
                     sync=True)
         if delete:
-            # if os.environ.get('DEV'):
-            #     az_path_base = f'staging/{user.id}/'
-            # else:
-            #     az_path_base = f'{user.id}/'
-
-            # az = boto3.resource('s3')
-            # buck = az.Bucket(bucket)
-            # buck.objects.filter(Prefix=az_path_base).delete()
-            delete_user(user)
-
-            # for tag in user.tags.all():
-            #     db.session.delete(tag)
-            # for highlight in user.highlights.all():
-            #     db.session.delete(highlight)
-            # for topic in user.topics.all():
-            #     db.session.delete(topic)
-            # for article in user.articles.all():
-            #     db.session.delete(article)
-            # _set_task_progress(100)
-            # db.session.delete(user)
-            # db.session.commit()
-        
+            delete_user(user.id)
+     
     except:
-        app.logger.error('Unhandled exception', exc_info=sys.exc_info())
+        logger.error('Unhandled exception', exc_info=sys.exc_info())
   
 
 
@@ -200,7 +183,7 @@ def export_highlights(user, highlights, source, ext):
             url = s3.generate_presigned_url('get_object', Params = {'Bucket': bucket, 'Key': az_path}, ExpiresIn = 604800)
 
 
-            print(f'sending email - [Lurnby] Your exported data for user: {user.id}')
+            logger.info(f'sending email - [Lurnby] Your exported data for user: {user.id}')
             send_email('[Lurnby] Your exported highlights',
                     sender=app.config['ADMINS'][0], recipients=[user.email],
                     text_body=render_template('email/export_highlights.txt', url=url, user=user),
@@ -240,7 +223,7 @@ def export_highlights(user, highlights, source, ext):
             # url = "https://s3-%s.amazonaws.com/%s/%s" % (location, bucket, az_path)
             url = s3.generate_presigned_url('get_object', Params = {'Bucket': bucket, 'Key': az_path}, ExpiresIn = 604800)
 
-            print(f'sending email - [Lurnby] Your exported data for user: {user.id}')
+            logger.info(f'sending email - [Lurnby] Your exported data for user: {user.id}')
             send_email('[Lurnby] Your exported highlights',
                     sender=app.config['ADMINS'][0], recipients=[user.email],
                     text_body=render_template('email/export_highlights.txt', url=url, user=user),
@@ -248,7 +231,7 @@ def export_highlights(user, highlights, source, ext):
                     sync=True)                
                 
     except:
-        app.logger.error('Unhandled exception', exc_info=sys.exc_info())
+        logger.error('Unhandled exception', exc_info=sys.exc_info())
     finally:
         _set_task_progress(100)
 
@@ -340,7 +323,7 @@ def bg_add_article(u, a_id, pdf, epub, tags):
             os.remove(path)
             
     except:
-        app.logger.error('Unhandled exception', exc_info=sys.exc_info())
+        logger.error('Unhandled exception', exc_info=sys.exc_info())
     finally:
         _set_task_progress(100)
 
@@ -355,7 +338,7 @@ def set_images_lazy(aid):
         a.content = str(soup.prettify())
         db.session.commit()
     except:
-        app.logger.error('Unhandled exception', exc_info=sys.exc_info())
+        logger.error('Unhandled exception', exc_info=sys.exc_info())
     finally:
         _set_task_progress(100)
 
@@ -371,18 +354,18 @@ def set_absolute_urls(aid):
                     if 'http' not in img['src']:
                         img['src'] = f'{a.source_url}{img["src"]}'
                 except:
-                    print('no src in image')
+                    logger.info('no src in image')
             links = soup.find_all('a')
             for l in links:
                 try:
                     if 'http' not in l['href']:
                         l['href'] = f'{a.source_url}{l["href"]}'
                 except:
-                    print('no href in url')
+                    logger.info('no href in url')
             a.content = str(soup.prettify())
             db.session.commit()
     except:
-        app.logger.error('Unhandled exception', exc_info=sys.exc_info())
+        logger.error('Unhandled exception', exc_info=sys.exc_info())
     finally:
         _set_task_progress(100)
   
@@ -396,7 +379,7 @@ def create_recall_text(highlightId):
         if (len(words) > 3):
             for i in range(0,len(words) // 3):
                 num = randint(0, len(words) -1)
-                words[num] = re.sub('[\w\d]+','_____', words[num])
+                words[num] = re.sub(r'[\w\d]+','_____', words[num])
         text.replace_with(' '.join(words))
         
     highlight.prompt = soup.prettify()
